oci: move cap UT to _linux_test.go
No substantial code change
